+++ src/main/java/net/thumbtack/sharding/core/map/KeyMapper.java
@@ -10,5 +10,5 @@ public interface KeyMapper {
      * @param key The key.
      * @return The shard id.
      */
-    public long shard(long key);
+    public int shard(long key);
 }
